requirements.txt
Django==5.2.10
channels==4.1.0
channels-redis==4.2.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-ratelimit==4.1.0
redis==5.2.1
psycopg2-binary==2.9.10
